+++ trabalho/sistema_bancario/Conta.java
@@ -2,9 +2,9 @@ package trabalho.sistema_bancario;
 
 public class Conta {
     public int id_conta;
-    public float saldo;
+    public double saldo;
 
-    public Conta(int id_conta,float saldo){
+    public Conta(int id_conta,double saldo){
         this.id_conta = id_conta;
         this.saldo = saldo;
     }
